+++ ghost/email-service/lib/helpers/register-helpers.js
